encapsulating permissions related activities
