+++ packages/react-native-sdk/src/providers/MediaDevices.tsx
@@ -4,9 +4,8 @@ import { MediaDeviceInfo, useStreamVideoStoreSetState } from '../contexts';
 import {
   isCameraPermissionGranted$,
   isMicPermissionGranted$,
-} from '../utils/StreamVideoRN/rxSubjects';
-import { concatMap } from 'rxjs/operators';
-import { BehaviorSubject, EMPTY, Observable } from 'rxjs';
+  subscribeToDevicesWhenPermissionGranted,
+} from '../utils/StreamVideoRN/permissions';
 
 /**
  * A renderless component that provides the audio and video devices to the store
@@ -15,27 +14,6 @@ import { BehaviorSubject, EMPTY, Observable } from 'rxjs';
  *
  * @category Device Management
  */
-
-const subscribeToDevicesWhenPermissionGranted = (
-  isDevicePermissionGranted$: BehaviorSubject<boolean>,
-  getDevicesFunc: () => Observable<MediaDeviceInfo[]>,
-  subscriptionCallback: (videoDevices: MediaDeviceInfo[]) => void,
-) =>
-  isDevicePermissionGranted$
-    .pipe(
-      concatMap((isDevicePermissionGranted) => {
-        // if we don't have mic permission, we don't need to get the audio devices
-        // because we won't be able to use them anyway and this will trigger a permission request
-        // from RN WebRTC lib. This is not ideal because we want to control when the permission.
-        if (!isDevicePermissionGranted) {
-          // otherwise return EMPTY, which is an Observable that does nothing and just completes immediately
-          return EMPTY;
-        }
-        return getDevicesFunc();
-      }),
-    )
-    .subscribe(subscriptionCallback);
-
 export const MediaDevices = (): React.ReactElement | null => {
   const setState = useStreamVideoStoreSetState();
   const initialVideoDeviceSet = useRef(false);
