--- packages/react-native-sdk/src/utils/StreamVideoRN/rxSubjects.ts
@@ -1,4 +1,0 @@
-import { BehaviorSubject } from 'rxjs';
-
-export const isCameraPermissionGranted$ = new BehaviorSubject<boolean>(false);
-export const isMicPermissionGranted$ = new BehaviorSubject<boolean>(false);
