+++ sample-apps/react-native/dogfood/App.tsx
@@ -28,6 +28,7 @@ import { setPushConfig } from './src/utils/setPushConfig';
 import { PermissionsAndroid, Platform } from 'react-native';
 import { StreamVideoRN } from '@stream-io/video-react-native-sdk';
 import { useAppStateListener } from 'stream-chat-react-native';
+import { useSyncPermissions } from './src/hooks/useSyncPermissions';
 
 // @ts-expect-error
 Logger.enable(false);
@@ -38,42 +39,6 @@ const Stack = createNativeStackNavigator<RootStackParamList>();
 // since the app can be opened from a dead state through a push notification
 setPushConfig();
 
-const checkAndUpdatePermissions = async () => {
-  // TODO(SG): ask with RN permissions
-  if (Platform.OS === 'ios') {
-    StreamVideoRN.setPermissions({
-      isCameraPermissionGranted: true,
-      isMicPermissionGranted: true,
-    });
-    return;
-  }
-  const results = await PermissionsAndroid.requestMultiple([
-    PermissionsAndroid.PERMISSIONS.RECORD_AUDIO,
-    PermissionsAndroid.PERMISSIONS.CAMERA,
-    PermissionsAndroid.PERMISSIONS.BLUETOOTH_CONNECT,
-    PermissionsAndroid.PERMISSIONS.POST_NOTIFICATIONS,
-  ]);
-  StreamVideoRN.setPermissions({
-    isCameraPermissionGranted:
-      results[PermissionsAndroid.PERMISSIONS.CAMERA] ===
-      PermissionsAndroid.RESULTS.GRANTED,
-    isMicPermissionGranted:
-      results[PermissionsAndroid.PERMISSIONS.RECORD_AUDIO] ===
-      PermissionsAndroid.RESULTS.GRANTED,
-  });
-};
-const handleOnForeground = async () => {
-  const isCameraPermissionGranted = await PermissionsAndroid.check(
-    PermissionsAndroid.PERMISSIONS.CAMERA,
-  );
-  const isMicPermissionGranted = await PermissionsAndroid.check(
-    PermissionsAndroid.PERMISSIONS.RECORD_AUDIO,
-  );
-  StreamVideoRN.setPermissions({
-    isCameraPermissionGranted,
-    isMicPermissionGranted,
-  });
-};
 const StackNavigator = () => {
   const appMode = useAppGlobalStoreValue((store) => store.appMode);
   const username = useAppGlobalStoreValue((store) => store.username);
@@ -81,14 +46,7 @@ const StackNavigator = () => {
   const setState = useAppGlobalStoreSetState();
 
   useProntoLinkEffect();
-
-  // Ask for relevant permissions after screen is loaded and when app state changes
-  // to foreground to make sure we have the latest permissions in order to subscribe
-  // to audio/video devices
-  useEffect(() => {
-    checkAndUpdatePermissions();
-  }, []);
-  useAppStateListener(handleOnForeground, () => {});
+  useSyncPermissions();
 
   let mode;
   switch (appMode) {
